requirements.txt
google-generativeai==0.3.2
streamlit==1.28.2
python-dotenv
PyPDF2
#langchain
pinecone-client
langchain==0.0.225
flask
langchain_google_genai
langchain-google-vertexai
langchain_core
docx
-e .